Move CountDown & Future-/MultiActionListener to io.crate.concurrent
MultiActionListener, and CountDown fill a similar use-case with
CountdownFuture, so this groups the classes in the same package.
