+++ server/src/main/java/io/crate/execution/engine/collect/collectors/NodeStats.java
@@ -36,9 +36,10 @@ import org.elasticsearch.cluster.node.DiscoveryNode;
 import org.elasticsearch.transport.ConnectTransportException;
 import org.elasticsearch.transport.ReceiveTimeoutTransportException;
 
-import io.crate.action.FutureActionListener;
 import io.crate.common.exceptions.Exceptions;
 import io.crate.common.unit.TimeValue;
+import io.crate.concurrent.FutureActionListener;
+import io.crate.concurrent.MultiActionListener;
 import io.crate.data.BatchIterator;
 import io.crate.data.CollectingBatchIterator;
 import io.crate.data.Row;
@@ -48,7 +49,6 @@ import io.crate.execution.engine.collect.RowsTransformer;
 import io.crate.execution.engine.collect.stats.NodeStatsRequest;
 import io.crate.execution.engine.collect.stats.NodeStatsResponse;
 import io.crate.execution.support.ActionExecutor;
-import io.crate.execution.support.MultiActionListener;
 import io.crate.expression.InputFactory;
 import io.crate.expression.reference.StaticTableReferenceResolver;
 import io.crate.expression.reference.sys.node.NodeStatsContext;
