+++ server/src/main/java/io/crate/execution/jobs/transport/TransportCancelAction.java
@@ -35,7 +35,7 @@ import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportResponse;
 import org.elasticsearch.transport.TransportService;
 
-import io.crate.execution.support.MultiActionListener;
+import io.crate.concurrent.MultiActionListener;
 import io.crate.execution.support.NodeAction;
 import io.crate.execution.support.NodeActionRequestHandler;
 import io.crate.session.Sessions;
