+++ server/src/main/java/io/crate/replication/logical/repository/LogicalReplicationRepository.java
@@ -73,10 +73,10 @@ import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.RemoteClusters;
 import org.jetbrains.annotations.Nullable;
 
-import io.crate.action.FutureActionListener;
 import io.crate.common.exceptions.Exceptions;
 import io.crate.common.io.IOUtils;
 import io.crate.common.unit.TimeValue;
+import io.crate.concurrent.FutureActionListener;
 import io.crate.exceptions.SQLExceptions;
 import io.crate.metadata.IndexName;
 import io.crate.metadata.RelationName;
