+++ server/src/test/java/org/elasticsearch/index/shard/IndexShardTests.java
@@ -147,10 +147,10 @@ import org.junit.Test;
 
 import com.carrotsearch.randomizedtesting.RandomizedTest;
 
-import io.crate.action.FutureActionListener;
 import io.crate.common.collections.Tuple;
 import io.crate.common.exceptions.Exceptions;
 import io.crate.common.unit.TimeValue;
+import io.crate.concurrent.FutureActionListener;
 import io.crate.metadata.doc.SysColumns;
 
 /**
