+++ server/src/test/java/org/elasticsearch/transport/TransportHandshakerTests.java
@@ -39,8 +39,8 @@ import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.threadpool.TestThreadPool;
 import org.junit.Test;
 
-import io.crate.action.FutureActionListener;
 import io.crate.common.unit.TimeValue;
+import io.crate.concurrent.FutureActionListener;
 import io.netty.channel.embedded.EmbeddedChannel;
 
 public class TransportHandshakerTests extends ESTestCase {
